refactor: AutoSyncConfig 엔티티의 패키지 위치 오류 수정
