+++ src/main/java/org/codeiteam3/findex/syncjob/service/SyncJobService.java
@@ -1,7 +1,7 @@
 package org.codeiteam3.findex.syncjob.service;
 
 import lombok.RequiredArgsConstructor;
-import org.codeiteam3.findex.autosyncconfig.AutoSyncConfig;
+import org.codeiteam3.findex.autosyncconfig.entity.AutoSyncConfig;
 import org.codeiteam3.findex.autosyncconfig.repository.AutoSyncConfigRepository;
 import org.codeiteam3.findex.pagination.CursorPageResponse;
 import org.codeiteam3.findex.pagination.CursorPageResponseMapper;
